# aibots-api-develop/openapi/full.yaml
openapi: 3.0.3
info:
  title: AIBots Full API
  description: This document describes the Full API interface for AIBots
  version: 1.0.0
  contact:
    name: Moonshot Team
    url: 'https://www.tech.gov.sg/capability-centre-dsaid'
    email: david_tw_lee@tech.gov.sg
servers:
  - url: 'https://aibots.gov.sg/v{major}.{minor}'
    description: Full server address hosted on cloud
    variables:
      major:
        enum:
          - "1"
        default: "1"
      minor:
        enum:
          - "0"
        default: "0"
...
paths:
...
  /definitions/agents:
    get:
      tags:
        - Schemas
      summary: Default settings form parameters for AI Agents
      description: >-
        Retrieves the default form parameters available for configuring AI 
        Agents
      operationId: get_agent_definitions
      responses:
        '200':
          description: Successfully retrieved Agent settings
          content:
            application/json:
              schema:
                description: Default Agent settings definitions
                type: array
                items:
                  $ref: '#/components/schemas/AtlasParam'
                default: []
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Agent APIs
  /agents:
    post:
      tags:
        - Agents
      summary: Creates a new Agent
      description: |-
        An Agent encapsulates various forms of interacting with a Large Language 
        Model (LLM), this could be as diverse as interacting with an image 
        generation model (multi-modal Chat Agent), to performing question and 
        answer on a specified knowledge base (RAG Agent), to performing actions
        on a user's behalf (AI Agent). An Agent can then be served out for other 
        users to interact with it. Agent are highly configurable, this provides 
        Agent creators with the flexibility to serve out Agents that address the 
        needs of their target audience. Agent access can be restricted to 
        different individuals or organisations/groups. They can be cloned with 
        the entire Agent configuration and customised according to their needs. 
        This API call creates a new Agent configuration that structures the 
        interaction.
      operationId: create_agent
      requestBody:
        description: Configuration of Agent to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPOST'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Agents
      summary: Retrieves all Agents visible to the user
      description: >-
        Returns a list of all Agents that the user is permitted to view and 
        invoke. Supports various query parameters for retrieving public and 
        shared Agents.
      operationId: get_agents
      parameters:
        - $ref: '#/components/parameters/IDListParam'
        - name: featured
          description: Retrive all featured Agents
          in: query
          required: false
          schema:
            type: boolean
        - name: public
          description: Retrieve all Agents available to the WOG
          in: query
          required: false
          schema:
            type: boolean
        - name: owner
          description: Retrieve all Agents that the user owns
          in: query
          required: false
          schema:
            type: boolean
        - name: admin
          description: Retrieve all Agents that the user can edit/manage
          in: query
          required: false
          schema:
            type: boolean
        - name: shared
          description: Retrieve all Agents shared with the user
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: Successfully retrieved Agents
          content:
            application/json:
              schema:
                description: List of Agents
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
                default: []
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  '/agents/{id}':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    put:
      tags:
        - Agents
      summary: Updates the core configuration of an existing Agent
      description: |-
        This API enables the user to update the main configuration of an 
        existing Agent.
      operationId: update_agent
      requestBody:
        description: Configuration of Agent to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentPUT'
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Agents
      summary: Retrieves the configuration of an existing Agent
      description: Retrieves an existing Agent
      operationId: get_agent
      responses:
        '200':
          description: Successfully retrieved Agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Agents
      summary: Deletes an existing Agent configuration
      description: |-
        Includes the knowledgeBase boolean flag to delete the Knowledge Base 
        associated with the Agent. If the Knowledge Base is shared by other Agents,
        then the files boolean flag will not apply. A push notification will be
        sent upon completion.
      operationId: delete_agent
      parameters:
        - description: Deletes all files associated with the Knowledge Bases
          name: files
          in: query
          required: false
          schema:
            type: boolean
      responses:
        '202':
          description: |-
            Successfully deleted Agent and associated Knowledge Base will be 
            cleared as a background task with a push notification sent.
        '204':
          $ref: '#/components/responses/DeletedResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/InvalidDeletionError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  '/agents/{id}/knowledge/bases/{secondary_id}/statuses':
    parameters:
      - $ref: '#/components/parameters/IDParam'
      - $ref: '#/components/parameters/SecondaryIDParam'
    put:
      tags:
        - Agents
      summary: Updates the current state of embeddings generated
      description: >-
        This API enables the user to update the Knowledge Base embeddings for 
        a given RAG config. The secondary_id parameter is the ID of the 
        KnowledgeBase while the ragConfig paramter indicates the RAGConfig ID
        to be updated
      operationId: update_embeddings
      parameters:
        - description: ID of the RAG Config to be updated
          name: ragConfig
          required: true
          schema:
            type: string
            format: uuid
          in: query
      requestBody:
        description: Current state of embeddings
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/State'
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/agents/{id}/rags':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    post:
      tags:
        - Agents
      summary: Adds new RAG pipeline configurations to an existing Agent
      description: >-
        This API enables the user to add the RAG pipeline configurations based
        on the supported RAG pipelines available. Users are able to add and 
        remove mutiple pipelines. After adding new RAG pipeline configurations, 
        users will need to trigger the new pipelines using the POST 
        /rags/{id}/pipelines endpoint to generate embeddings. As each Agent can 
        support multiple pipelines, these pipelines will be queried in the order
        provided and function as failovers in the event that one RAG service 
        goes down. There are several approaches to switching pipelines:
        
          1. By failing over. The failover feature also enables users to switch 
             pipelines by first adding a new pipeline before the existing one 
             and triggering it. While the new pipeline is running, the existing 
             queries will fail over to the existing one, allowing the Agent to 
             maintain functionality. When the new pipeline is complete, queries 
             will automatically be served from the new pipeline as it is 
             triggered first. The old pipeline can be cleared subsequently.
          2. In sequential order. As pipelines are queried in sequential order, 
             a user can add a new pipeline configuration after the current and 
             trigger it to run. However, the user must explicitly clear the 
             original pipeline in order to perform the switch.
      operationId: create_rag_configs
      requestBody:
        description: RAG pipeline configuration to be added
        required: true
        content:
          application/json:
            schema:
              description: List of supported RAG pipelines
              type: array
              items:
                $ref: '#/components/schemas/RAGConfig'
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Agents
      summary: Retrieves all RAG pipeline configurations associated with the Agent
      description: |-
        Returns a list of all RAG pipeline configurations associated with the Agent
      operationId: get_rag_configs
      responses:
        '200':
          description: Successfully retrieved all RAG pipeline configurations belonging to an Agent
          content:
            application/json:
              schema:
                description: List of RAG pipeline configurations
                type: array
                items:
                  $ref: '#/components/schemas/RAGConfig'
                default: []
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Agents
      summary: Deletes existing RAG pipeline configurations from an existing Agent
      description: >-
        This API enables the user to delete existing RAG pipeline configurations.
        When a RAG pipeline configuration is deleted, embeddings generated from
        the associated RAG pipeline are cleared.
      operationId: delete_rag_configs
      parameters:
        - $ref: '#/components/parameters/IDListParam'
      responses:
        '204':
          $ref: '#/components/responses/DeletedResponse'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  '/agents/{id}/rags/{secondary_id}/statuses':
    parameters:
      - $ref: '#/components/parameters/IDParam'
      - $ref: '#/components/parameters/SecondaryIDParam'
    put:
      tags:
        - Agents
      summary: Updates the current state of the RAG config
      description: >-
        This API enables the user to update the RAG config state
      operationId: update_rag_config
      requestBody:
        description: Current state of the RAG config
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/State'
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  '/rags/{id}/pipelines/status':
    parameters:
      - $ref: '#/components/parameters/IDParam'
    get:
      tags:
        - RAG
      security:
        - AtlasAPIKey: []
      summary: Retrieves the latest pipeline status
      description: >-
        Allows users to retrieve the latest RAG pipeline run configuration 
        status of an Agent
      operationId: get_rag_pipeline_status
      responses:
        '204':
          $ref: '#/components/responses/UpdatedResponse'
        '400':
          $ref: '#/components/responses/InvalidParametersError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '403':
          $ref: '#/components/responses/PermissionsError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Notifications
...
security:
  - AtlasAuth: []
components:
  schemas:
...
    AtlasParam:
      description: Representation of a dynamic configurable parameter field
      type: object
      properties:
        name:
          description: Display name of the Parameter
          type: string
        description:
          description: Brief description of the Parameter
          type: string
          default: ''
        key:
          description: Internal key of the Parameter
          type: string
        default:
          description: Default parameter value
          nullable: true
          default: null
        required:
          description: Indicates if the Parameter is required
          type: boolean
          default: false
        schema:
          description: Json schema structure to validate the Parameter against
          type: object
          additionalProperties: true
          default: {}
        tags:
          description: Tags associated with the Parameter
          type: array
          items:
            type: string
          default: []
        componentConfig:
          description: UI components to render the Param configuration form with
          anyOf:
            - $ref: '#/components/schemas/AtlasComponents'
            - description: List of Atlas Components
              type: array
              items:
                $ref: '#/components/schemas/AtlasComponents'
          nullable: true
          default: null
      required:
        - name
        - description
        - key
        - default
        - schema
        - required
        - tags
        - componentConfig

    # States
...
    Agent:
      description: Representation of a Agent
      type: object
      properties:
        id:
          description: Unique ID of the Agent
          type: string
          format: uuid
          readOnly: true
        name:
          description: Name of the Agent
          type: string
        description:
          description: Brief description of the Agent
          type: string
        featured:
          description: Indicates if the Agent is a featured Agent
          type: boolean
          readOnly: true
        clone:
          description: ID of the original Agent cloned from
          type: string
          format: uuid
          readOnly: true
          nullable: true
          default: null
        sharing:
          $ref: '#/components/schemas/AgentSharing'
        ownership:
          $ref: '#/components/schemas/Ownership'
        files:
          description: Non-embedded files associated with the Agent
          type: array
          items:
            type: string
            format: uuid
          default: []
        agency:
          description: Agency associated with the Agent
          type: string
        welcomeMessage:
          description: Welcome message the Agent will display when a new chat is created
          type: string
          default: "Hello! I am your assistant. How can I help you?"
        templates:
          description: Default templates to use with the Agent
          type: array
          items:
            $ref: '#/components/schemas/AgentTemplate'
          default: []
        releaseState:
          $ref: '#/components/schemas/AgentReleaseState'
        knowledgeBases:
          description: Knowledge Bases associated with Agent
          type: array
          items:
            type: string
            format: uuid
          default: []
        defaultPipeline:
          description: Default RAG pipeline to be used
          type: string
          nullable: true
          default: null
        rags:
          description: List of supported RAG pipelines
          type: array
          items:
            type: string
            format: uuid
          default: []
        chat:
          $ref: '#/components/schemas/AgentChatConfig'
        tools:
          description: Tools that the Agent is able to access
          type: array
          items:
            type: string
            format: uuid
          default: []
        settings:
          description: >-
            Additional settings associated with the Agent
          type: object
          additionalProperties: true
          default: {}
          example:
            conversationHistory: true
            accessDocs: true
            citations: true
            userCreativity: true
            userMaxLength: true
            userRemovePII: true
            userCustomInstructions: true
        tags:
          description: Tags associated with the Agent
          type: array
          items:
            type: string
          default: []
        meta:
          $ref: '#/components/schemas/Meta'
      required:
        - id
        - name
        - description
        - clone
        - featured
        - sharing
        - ownership
        - files
        - agency
        - welcomeMessage
        - templates
        - releaseState
        - knowledgeBases
        - defaultPipeline
        - rags
        - chat
        - tools
        - settings
        - tags
        - meta
...
    AgentPOST:
      description: POST request body for creating a Agent
      type: object
      properties:
        name:
          description: Name of the Agent
          type: string
        description:
          description: Brief description of the Agent
          type: string
        ownership:
          $ref: '#/components/schemas/Ownership'
        files:
          description: Non-embedded files associated with the Agent
          type: array
          items:
            type: string
            format: uuid
          default: []
        welcomeMessage:
          description: Welcome message the Agent will display when a new chat is created
          type: string
          default: "Hello! I am your assistant. How can I help you?"
        agency:
          description: Agency associated with the Agent
          type: string
        templates:
          description: Default templates to use with the Agent
          type: array
          items:
            $ref: '#/components/schemas/AgentTemplate'
          default: []
        knowledgeBases:
          description: Knowledge bases associated with Agent
          type: array
          items:
            $ref: '#/components/schemas/DataSource'
          default: []
        defaultPipeline:
          description: Default RAG pipeline to be used
          type: string
          nullable: true
          default: null
        rags:
          description: List of supported RAG pipelines
          type: array
          items:
            $ref: '#/components/schemas/RAGConfig'
          default: []
        chat:
          $ref: '#/components/schemas/AgentChatConfig'
        tools:
          description: Tools that the Agent is able to access
          type: array
          items:
            type: string
            format: uuid
          default: []
        settings:
          description: >-
            Additional settings associated with the Agent
          type: object
          additionalProperties: true
          default: {}
          example:
            conversationHistory: true
            accessDocs: true
            citations: true
            userCreativity: true
            userMaxLength: true
            userRemovePII: true
            userCustomInstructions: true
      required:
        - name
        - agency
    AgentPUT:
      description: PUT request body for updating an Agent
      type: object
      properties:
        name:
          description: Name of the Agent
          type: string
        description:
          description: Brief description of the Agent
          type: string
        files:
          description: Non-embedded files associated with the Agent
          type: array
          items:
            type: string
            format: uuid
        agency:
          description: Agency associated with the Agent
          type: string
        welcomeMessage:
          description: Welcome message the Agent will display when a new chat is created
          type: string
        templates:
          description: Default templates to use with the Agent
          type: array
          items:
            $ref: '#/components/schemas/AgentTemplate'
        chat:
          $ref: '#/components/schemas/AgentChatConfig'
        tools:
          description: Tools that the Agent is able to access
          type: array
          items:
            type: string
            format: uuid
          default: []
        settings:
          description: >-
            Additional settings associated with the Agent
          type: object
          additionalProperties: true
...
  parameters:
    IDParam:
      name: id
      description: A unique ID string
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IDListParam:
      name: ids
      description: A list of optional IDs
      in: query
      schema:
        type: array
        items:
          type: string
          format: uuid
      style: form
      explode: true
...
  responses:
    CreatedResponse:
      description: Successfully created resource with ID
      content:
        application/json:
          schema:
            description: Generated resource ID
            type: object
            properties:
              id:
                description: Unique ID of the resource
                type: string
                format: uuid
            required:
              - id
    UpdatedResponse:
      description: Successfully updated specified resource
...
    InvalidParametersError:
      description: Invalid parameter(s)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 400
            message: Invalid parameters provided
    AuthenticationError:
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 401
            message: Unauthorised usage of the API
    PermissionsError:
      description: Insufficient permissions error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 403
            message: Insufficient permissions to use API
    NotFoundError:
      description: Specified resource(s) was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 404
            message: Specified resource(s) was not found
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
          example:
            code: 500
            message: Internal server error

    # SCIM responses
...
  securitySchemes:
    AtlasAuth:
      description: Atlas JWT authentication token
      type: http
      scheme: bearer
      bearerFormat: JWT
    AtlasAPIKey:
      description: Atlas API Key authentication
      type: apiKey
      in: header
      name: X-ATLAS-Key